Remove TradeContent container
